src/agents/mcp/security/__init__.py
"""
Adaptive Security Mesh for MCP

This module provides a comprehensive security framework for the MCP system,
implementing adaptive security measures based on context and threat intelligence.
"""

from .types import (
    SecurityLevel,
    AccessLevel,
    SecurityPolicy,
    SecurityConfig
)
from .features import SecurityFeatures
from .mesh import SecurityMesh, ProtocolManager, CredentialManager
from .integration import SecurityMiddleware

__all__ = [
    'SecurityLevel',
    'AccessLevel',
    'SecurityPolicy',
    'SecurityConfig',
    'SecurityFeatures',
    'SecurityMesh',
    'ProtocolManager',
    'CredentialManager',
    'SecurityMiddleware',
    'load_security_policy'
]

from typing import Dict, List, Optional, Union, Any
from dataclasses import dataclass
from enum import Enum
from pathlib import Path
import yaml
from .features import SecurityFeatures, AccessLevel

class SecurityLevel(Enum):
    LOW = 1
    MEDIUM = 2
    HIGH = 3
    CRITICAL = 4

class AccessLevel(Enum):
    PUBLIC = 1
    INTERNAL = 2
    CONFIDENTIAL = 3
    RESTRICTED = 4

@dataclass
class SecurityPolicy:
    """Configuration for security mesh behavior"""
    default_level: SecurityLevel
    sensitivity_thresholds: Dict[str, float]
    protocol_mappings: Dict[SecurityLevel, List[str]]
    credential_requirements: Dict[SecurityLevel, List[str]]
    threat_intelligence_sources: List[str]

@dataclass
class SecurityConfig:
    """Security configuration container"""
    policy_file: Path
    threat_intelligence_sources: List[str]
    credential_stores: List[str]
    protocol_registry: Dict[str, Dict]

    @classmethod
    def from_yaml(cls, config_path: Path) -> 'SecurityConfig':
        """Load configuration from YAML file"""
        with open(config_path) as f:
            config_data = yaml.safe_load(f)
        
        return cls(
            policy_file=Path(config_data['policy_file']),
            threat_intelligence_sources=config_data['threat_intelligence_sources'],
            credential_stores=config_data['credential_stores'],
            protocol_registry=config_data['protocol_registry']
        )

class SecurityMesh:
    """Core security mesh implementation"""
    def __init__(self, policy: SecurityPolicy):
        self.policy = policy
        self.classification_engine = ClassificationEngine()
        self.protocol_manager = ProtocolManager()
        self.credential_manager = CredentialManager()
        self.threat_intelligence = ThreatIntelligenceManager()
        self.security_features = SecurityFeatures(policy)

    async def validate_server_identity(self, server: Any) -> None:
        """Validate server identity and security posture"""
        # Implementation would verify server certificates, security configuration, etc.
        pass

    async def filter_tools(self, tools: List[Any]) -> List[Any]:
        """Filter tools based on security policy"""
        # Implementation would filter tools based on access control
        return tools

    async def validate_tool_access(self, tool_name: str) -> None:
        """Validate tool access permissions"""
        user_context = {
            "access_level": AccessLevel.INTERNAL,  # This would come from actual user context
            "roles": ["user"],  # This would come from actual user context
            "credentials": ["basic_auth"]  # This would come from actual user context
        }
        
        if not await self.security_features.validate_tool_access(tool_name, user_context):
            raise SecurityError(f"Access denied to tool: {tool_name}")

    async def process_input(self, arguments: Dict[str, Any]) -> Dict[str, Any]:
        """Process and secure input arguments"""
        return await self.security_features.process_input("tool_name", arguments)

    async def process_output(self, result: Any) -> Any:
        """Process and secure output result"""
        return await self.security_features.process_output("tool_name", result)

    async def audit_cleanup(self, server: Any) -> None:
        """Audit server cleanup"""
        # Implementation would log security events
        pass

class SecurityError(Exception):
    """Security-related exception"""
    pass

class ClassificationEngine:
    """Data classification engine"""
    def __init__(self):
        pass

class ProtocolManager:
    """Dynamic cryptographic protocol management"""
    def __init__(self):
        self.available_protocols = {}
        self.active_protocol = None

    async def select_protocol(self, security_level: SecurityLevel) -> List[str]:
        """Select appropriate protocol based on security level"""
        return ["tls_1_3"]  # Placeholder implementation

class CredentialManager:
    """Automatic credential escalation management"""
    def __init__(self):
        self.credential_store = {}
        self.escalation_policies = {}

    async def get_credentials(self, security_level: SecurityLevel) -> List[str]:
        """Get required credentials for security level"""
        return ["basic_auth"]  # Placeholder implementation

class ThreatIntelligenceManager:
    """Threat intelligence integration"""
    def __init__(self):
        self.sources = []
        self.current_threat_level = 0.0

    async def get_current_level(self) -> float:
        """Get current threat level from all sources"""
        return 0.5  # Placeholder implementation

class SecurityFeatures:
    """Implementation of specific security features"""
    def __init__(self, policy: SecurityPolicy):
        self.policy = policy
        self.tool_policies: Dict[str, Dict] = {}
        self._load_tool_policies()

    def _load_tool_policies(self):
        """Load tool access policies from configuration"""
        self.tool_policies = {
            "read_data": {
                "required_level": AccessLevel.PUBLIC,
                "allowed_roles": {"user", "admin"},
                "required_credentials": ["basic_auth"],
                "input_sensitivity": {"query": SecurityLevel.LOW},
                "output_sensitivity": SecurityLevel.LOW
            },
            "process_sensitive": {
                "required_level": AccessLevel.CONFIDENTIAL,
                "allowed_roles": {"admin"},
                "required_credentials": ["oauth2", "certificate"],
                "input_sensitivity": {"data": SecurityLevel.HIGH},
                "output_sensitivity": SecurityLevel.HIGH
            }
        }

    async def validate_tool_access(
        self, 
        tool_name: str, 
        user_context: Dict[str, Any]
    ) -> bool:
        """Validate if user can access a specific tool"""
        if tool_name not in self.tool_policies:
            return False
            
        policy = self.tool_policies[tool_name]
        
        # Check access level
        if user_context.get("access_level", AccessLevel.PUBLIC) < policy["required_level"]:
            return False
            
        # Check roles
        user_roles = set(user_context.get("roles", []))
        if not user_roles.intersection(policy["allowed_roles"]):
            return False
            
        # Check credentials
        user_credentials = set(user_context.get("credentials", []))
        if not all(cred in user_credentials for cred in policy["required_credentials"]):
            return False
            
        return True

    async def process_input(
        self, 
        tool_name: str, 
        arguments: Dict[str, Any]
    ) -> Dict[str, Any]:
        """Process and secure input arguments"""
        if tool_name not in self.tool_policies:
            return arguments
            
        policy = self.tool_policies[tool_name]
        secured_args = {}
        
        for key, value in arguments.items():
            sensitivity = policy["input_sensitivity"].get(key, SecurityLevel.LOW)
            secured_args[key] = await self._secure_value(value, sensitivity)
            
        return secured_args

    async def process_output(
        self, 
        tool_name: str, 
        result: Any
    ) -> Any:
        """Process and secure output result"""
        if tool_name not in self.tool_policies:
            return result
            
        policy = self.tool_policies[tool_name]
        return await self._secure_value(result, policy["output_sensitivity"])

    async def _secure_value(
        self, 
        value: Any, 
        sensitivity: SecurityLevel
    ) -> Any:
        """Apply security measures based on sensitivity level"""
        if isinstance(value, str):
            if sensitivity == SecurityLevel.LOW:
                return value
            elif sensitivity == SecurityLevel.MEDIUM:
                return self._mask_sensitive_patterns(value)
            elif sensitivity == SecurityLevel.HIGH:
                return f"ENCRYPTED:{value}"
            elif sensitivity == SecurityLevel.CRITICAL:
                return f"CONTROLLED:ENCRYPTED:{value}"
        return value

    def _mask_sensitive_patterns(self, text: str) -> str:
        """Mask sensitive patterns in text"""
        import re
        patterns = {
            r'\b\d{16}\b': '****-****-****-****',  # Credit card
            r'\b\d{3}-\d{2}-\d{4}\b': '***-**-****',  # SSN
            r'\b[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Z|a-z]{2,}\b': '***@***.***'  # Email
        }
        
        for pattern, replacement in patterns.items():
            text = re.sub(pattern, replacement, text)
            
        return text 

def load_security_policy(config: SecurityConfig) -> dict:
    """Load security policy from configuration"""
    # For now, return a basic policy
    return {
        "tools": {
            "read_data": {
                "required_level": AccessLevel.INTERNAL,
                "required_roles": ["admin", "reader"],
                "required_credentials": ["oauth2"],
                "input_sensitivity": SecurityLevel.MEDIUM,
                "output_sensitivity": SecurityLevel.MEDIUM
            }
        },
        "protocols": {
            "tls_1_3": {"name": "TLS 1.3", "min_strength": 256},
            "aes_256": {"name": "AES-256", "mode": "GCM", "key_length": 256}
        }
    } 